server/requirements.txt
Flask==2.3.2              # Web framework
Flask-SQLAlchemy==3.0.5   # ORM for database management
Flask-Migrate==4.0.5      # Database migrations
Flask-CORS==4.0.0         # Cross-Origin Resource Sharing
Flask-JWT-Extended==4.5.2 # JWT authentication
Flask-Bcrypt==1.0.1       # Password hashing
Flask-Mail==0.9.1         # Email sending
python-dotenv==1.0.0      # Environment variable management
psycopg2-binary==2.9.6    # PostgreSQL adapter
google-auth==2.23.0       # Google OAuth for login
requests==2.31.0          # HTTP requests for Google/Apple login
PyJWT==2.8.0              # JWT handling for Apple login
stripe==7.0.0             # Stripe payment integration